arenasql: use protobuf to store table schema
  - if table schema format is changed, using bincode will break
    old tables; so, serialize table schema using protobuf so that
    updating table schema format doesnt make old table useless

+++ crates/arenasql/src/schema/column.rs
@@ -18,7 +18,7 @@ bitflags! {
     Serialize,
     Deserialize,
   )]
-  pub struct ColumnProperty: u16 {
+  pub struct ColumnProperty: u32 {
     const DEFAULT = 0;
     const ARCHIVED = 1 << 1;
     const NOT_NULL = 1 << 2;
